rename game -> game_specs
