+++ crate/src/game_specs/components/collision_pair.rs
