+++ crate/src/game_specs/event.rs
@@ -1,4 +1,4 @@
-use crate::game::controls::keyboard::Key;
+use crate::game_specs::controls::keyboard::Key;
 
 pub enum Event {
   MouseDown(i32, i32),
