+++ crate/src/game_specs/resources/collision_data.rs
